liberian-nlp-platform/backend/requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.11
celery==5.3.4
redis==5.0.1
boto3==1.29.7
Pillow==11.0.0
python-decouple==3.8
PyJWT==2.8.0
cryptography==41.0.7
langdetect==1.0.9
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2